+++ tests/TestCase.php
@@ -126,6 +126,9 @@ GRAPHQL;
             ]
         );
 
+        // TODO remove when the default changes
+        $config->set('lighthouse.force_fill', true);
+
         $config->set('app.debug', true);
     }