+++ CHANGELOG.md
@@ -9,6 +9,11 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
+### Added
+
+- Write definitions for programmatically registered types to `programmatic-types.graphql`
+  when running the `lighthouse:ide-helper` artisan command https://github.com/nuwave/lighthouse/pull/1371
+
 ## 4.13.1
 
 ### Fixed