+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 You can find and compare releases at the [GitHub release page](https://github.com/nuwave/lighthouse/releases).
 
 ## Unreleased
-
+- Added `@withCount` directive to allow eager loading relationship counts.
+- 
 ### Added
 
 - Write definitions for programmatically registered types to `programmatic-types.graphql`